Delete config add-image command
This no longer makes sense since replacement is performed using only JSON
paths.
