+++ pkg/commands/config.go
@@ -21,7 +21,6 @@ func NewConfigCommand() *cobra.Command {
 	}
 
 	cmd.AddCommand(
-		config.NewAddImageCommand(),
 		config.NewSetUserDefinedImage(),
 		config.NewDeleteUserDefinedImage(),
 		config.NewGetCommand(),
