+++ pkg/fs/bundle_config.go
@@ -8,10 +8,7 @@ package fs
 
 import (
 	"encoding/json"
-	"fmt"
 	"io"
-
-	"github.com/pivotal/image-relocation/pkg/images"
 
 	"github.com/bryanl/sheaf/pkg/sheaf"
 )
@@ -31,16 +28,10 @@ func NewBundleConfigCodec() *BundleConfigCodec {
 
 // Encode encodes a bundle config.
 func (b BundleConfigCodec) Encode(w io.Writer, bc sheaf.BundleConfig) error {
-	var imageList []string
-	if list := bc.GetImages(); list != nil {
-		imageList = list.Strings()
-	}
-
 	bcf := bundleConfigFormat{
 		SchemaVersion:     bc.GetSchemaVersion(),
 		Name:              bc.GetName(),
 		Version:           bc.GetVersion(),
-		Images:            imageList,
 		UserDefinedImages: bc.GetUserDefinedImages(),
 	}
 
@@ -56,16 +47,10 @@ func (b BundleConfigCodec) Decode(r io.Reader) (sheaf.BundleConfig, error) {
 		return nil, err
 	}
 
-	imageList, err := images.New(bcf.Images...)
-	if err != nil {
-		return nil, fmt.Errorf("images are invalid: %w", err)
-	}
-
 	bc := BundleConfig{
 		schemaVersion:     bcf.SchemaVersion,
 		name:              bcf.Name,
 		version:           bcf.Version,
-		images:            &imageList,
 		userDefinedImages: bcf.UserDefinedImages,
 	}
 
@@ -79,8 +64,6 @@ type bundleConfigFormat struct {
 	Name string `json:"name"`
 	// Version is the version of the fs.
 	Version string `json:"version"`
-	// Images is a set of images required by the fs.
-	Images []string `json:"images,omitempty"`
 	// UserDefinedImages is a list of user defined image locations.
 	UserDefinedImages []sheaf.UserDefinedImage `json:"userDefinedImages,omitempty"`
 }
@@ -93,8 +76,6 @@ type BundleConfig struct {
 	name string
 	// Version is the version of the fs.
 	version string
-	// Images is a set of images required by the fs.
-	images *images.Set
 	// UserDefinedImages is a list of user defined image locations.
 	userDefinedImages []sheaf.UserDefinedImage
 }
@@ -129,16 +110,6 @@ func (b BundleConfig) GetVersion() string {
 // SetVersion sets the bundle config's version.
 func (b *BundleConfig) SetVersion(version string) {
 	b.version = version
-}
-
-// GetImages returns the bundle config's images.
-func (b BundleConfig) GetImages() *images.Set {
-	return b.images
-}
-
-// SetImages sets the bundle config's images.
-func (b *BundleConfig) SetImages(imageSet *images.Set) {
-	b.images = imageSet
 }
 
 // GetUserDefinedImages returns the bundle config's user defined images.
