+++ pkg/fs/testdata/bundle-config-writer/bundle-no-images.json
@@ -6,7 +6,7 @@
     {
       "apiVersion": "v1",
       "kind": "Item",
-      "jsonPath": "{.}"
+      "jsonPath": "."
     }
   ]
 }
