+++ pkg/fs/testdata/bundle-config-writer/bundle.json
@@ -2,14 +2,11 @@
   "schemaVersion": "v1alpha1",
   "name": "test",
   "version": "0.1.0",
-  "images": [
-    "docker.io/library/foo"
-  ],
   "userDefinedImages": [
     {
       "apiVersion": "v1",
       "kind": "Item",
-      "jsonPath": "{.}"
+      "jsonPath": "."
     }
   ]
 }
